pyproject.toml
[project]
name = "company-crud-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Mike Shmygel",email = "m.shmygel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "django-debug-toolbar (>=5.1.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
pytest-cov = "^6.1.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "company_crud_api.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=. --cov-report=term-missing"

